First batch of changes. Printing assignments 'CommandLineOptions.Clo.PrintDesugarings = true;' are still failing and also calls that pass 'CommandLineOptions.Clo' where SMTLibOptions are expected
